add in_memory param in Python function

+++ function_for_download/get_nba_data.py
@@ -1,37 +1,50 @@
 from pathlib import Path
 from itertools import product
-import urllib.request
+from urllib.request import urlopen
 import tarfile
-from typing import Union, Sequence
+from typing import Union, Sequence, Optional, List
+from io import BytesIO, TextIOWrapper
+import csv
+
+import pandas as pd # If use in_memory = True & use_pandas = True
 
 
 def load_nba_data(path: Union[Path, str] = Path.cwd(),
-                  seasons: Union[Sequence, int] = range(1996, 2023),
+                  seasons: Union[Sequence, int] = range(1996, 2024),
                   data: Union[Sequence, str] = ("datanba", "nbastats", "pbpstats",
                                                 "shotdetail", "cdnnba", "nbastatsv3"),
                   seasontype: str = 'rg',
                   league: str = 'nba',
-                  untar: bool = False) -> None:
+                  untar: bool = False,
+                  in_memory: bool = False,
+                  use_pandas: bool = True) -> Optional[Union[List, pd.DataFrame]]:
     """
     Loading a nba play-by-play dataset from github repository https://github.com/shufinskiy/nba_data
 
     Args:
-        path (Union[Path, str]): Path where downloaded file should be saved on the hard disk
+        path (Union[Path, str]): Path where downloaded file should be saved on the hard disk. Not used if in_memory = True
         seasons (Union[Sequence, int]): Sequence or integer of the year of start of season
         data (Union[Sequence, str]): Sequence or string of data types to load
         seasontype (str): Part of season: rg - Regular Season, po - Playoffs
         league (str): Name league: NBA or WNBA
-        untar (bool): Logical: do need to untar loaded archive
+        untar (bool): Logical: do need to untar loaded archive. Not used if in_memory = True
+        in_memory (bool): Logical: If True dataset is loaded into workflow, without saving file to disk
+        use_pandas (bool): Logical: If True dataset is loaded how pd.DataFrame, else List[List[str]]. Ignore if in_memory=False
 
     Returns:
-        None
+        Optional[pd.DataFrame, List]: If in_memory=True and use_pandas=True return dataset how pd.DataFrame.
+        If use_pandas=False return dataset how List[List[str]]
+        If in_memory=False return None
     """
     if isinstance(path, str):
-        path = Path(path)
+        path = Path(path).expanduser()
     if isinstance(seasons, int):
         seasons = (seasons,)
     if isinstance(data, str):
         data = (data,)
+
+    if (len(data) > 1) & in_memory:
+        raise ValueError("Parameter in_memory=True available only when loading a single data type")
 
     if seasontype == 'rg':
         need_data = tuple(["_".join([data, str(season)]) for (data, season) in product(data, seasons)])
@@ -46,7 +59,12 @@ def load_nba_data(path: Union[Path, str] = Path.cwd(),
     if league.lower() == 'wnba':
         need_data = ['wnba_' + x for x in need_data]
 
-    with urllib.request.urlopen("https://raw.githubusercontent.com/shufinskiy/nba_data/main/list_data.txt") as f:
+    check_data = [file + ".csv" if untar else "tar.xz" for file in need_data]
+    not_exists = [not path.joinpath(check_file).is_file() for check_file in check_data]
+
+    need_data = [file for (file, not_exist) in zip(need_data, not_exists) if not_exist]
+
+    with urlopen("https://raw.githubusercontent.com/shufinskiy/nba_data/main/list_data.txt") as f:
         v = f.read().decode('utf-8').strip()
 
     name_v = [string.split("=")[0] for string in v.split("\n")]
@@ -55,12 +73,35 @@ def load_nba_data(path: Union[Path, str] = Path.cwd(),
     need_name = [name for name in name_v if name in need_data]
     need_element = [element for (name, element) in zip(name_v, element_v) if name in need_data]
 
+    if in_memory:
+        if use_pandas:
+            table = pd.DataFrame()
+        else:
+            table = []
     for i in range(len(need_name)):
-        t = urllib.request.urlopen(need_element[i])
-        with path.joinpath("".join([need_name[i], ".tar.xz"])).open(mode='wb') as f:
-            f.write(t.read())
-        if untar:
-            with tarfile.open(path.joinpath("".join([need_name[i], ".tar.xz"]))) as f:
-                f.extract("".join([need_name[i], ".csv"]), path)
+        with urlopen(need_element[i]) as response:
+            if response.status != 200:
+                raise Exception(f"Failed to download file: {response.status}")
+            file_content = response.read()
+            if in_memory:
+                with tarfile.open(fileobj=BytesIO(file_content), mode='r:xz') as tar:
+                    csv_file_name = "".join([need_name[i], ".csv"])
+                    csv_file = tar.extractfile(csv_file_name)
+                    if use_pandas:
+                        table = pd.concat([table, pd.read_csv(csv_file)], axis=0, ignore_index=True)
+                    else:
+                        csv_reader = csv.reader(TextIOWrapper(csv_file, encoding="utf-8"))
+                        for row in csv_reader:
+                            table.append(row)
+            else:
+                with path.joinpath("".join([need_name[i], ".tar.xz"])).open(mode='wb') as f:
+                    f.write(file_content)
+                if untar:
+                    with tarfile.open(path.joinpath("".join([need_name[i], ".tar.xz"]))) as f:
+                        f.extract("".join([need_name[i], ".csv"]), path)
 
-            path.joinpath("".join([need_name[i], ".tar.xz"])).unlink()
+                    path.joinpath("".join([need_name[i], ".tar.xz"])).unlink()
+    if in_memory:
+        return table
+    else:
+        return None
